# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Legacy Doc API",
    "description": "Documentacao automatica multi-linguagem com orquestracao multi-agente e multi-provedor.",
    "version": "2.0.0"
  },
  "paths": {
...
    "/v1/admin/accounts": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List Accounts",
        "description": "Contas com o que decide uma acao: gasto contra o teto, uso e situacao.",
        "operationId": "list_accounts_v1_admin_accounts_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Trecho do e-mail",
              "title": "Search"
            },
            "description": "Trecho do e-mail"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminAccountList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/accounts/{user_id}/plan": {
      "patch": {
        "tags": [
          "admin"
        ],
        "summary": "Change Plan",
        "description": "Troca o plano da conta, exigindo um motivo escrito.",
        "operationId": "change_plan_v1_admin_accounts__user_id__plan_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminPlanChangeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminAccountResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/accounts/{user_id}/status": {
      "patch": {
        "tags": [
          "admin"
        ],
        "summary": "Change Status",
        "description": "Ativa ou desativa a conta.",
        "operationId": "change_status_v1_admin_accounts__user_id__status_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminStatusChangeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminAccountResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/audit": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List Audit",
        "description": "Historico de acoes, da mais recente para a mais antiga.",
        "operationId": "list_audit_v1_admin_audit_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminActionResponse"
                  },
                  "title": "Response List Audit V1 Admin Audit Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AdminAccountList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AdminAccountResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "AdminAccountList"
      },
      "AdminAccountResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "plan_tier": {
            "type": "string",
            "title": "Plan Tier"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "is_admin": {
            "type": "boolean",
            "title": "Is Admin"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "jobs_this_month": {
            "type": "integer",
            "title": "Jobs This Month"
          },
          "spent_this_month_usd": {
            "type": "number",
            "title": "Spent This Month Usd"
          },
          "cost_limit_usd": {
            "type": "number",
            "title": "Cost Limit Usd",
            "description": "Teto do plano, para comparar com o gasto."
          },
          "documents_total": {
            "type": "integer",
            "title": "Documents Total"
          },
          "last_job_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Job At"
          },
          "locked_until": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locked Until",
            "description": "Preenchido enquanto o bloqueio por tentativas estiver ativo."
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "display_name",
          "plan_tier",
          "is_active",
          "is_admin",
          "created_at",
          "jobs_this_month",
          "spent_this_month_usd",
          "cost_limit_usd",
          "documents_total",
          "last_job_at"
        ],
        "title": "AdminAccountResponse",
        "description": "Uma conta vista pelo painel."
      },
      "AdminActionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "actor_email": {
            "type": "string",
            "title": "Actor Email"
          },
          "target_email": {
            "type": "string",
            "title": "Target Email"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "value_before": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value Before"
          },
          "value_after": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value After"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "actor_email",
          "target_email",
          "action",
          "value_before",
          "value_after",
          "reason",
          "created_at"
        ],
        "title": "AdminActionResponse"
      },
      "AdminPlanChangeRequest": {
        "properties": {
          "plan": {
            "type": "string",
            "title": "Plan"
          },
          "reason": {
            "type": "string",
            "maxLength": 500,
            "minLength": 5,
            "title": "Reason",
            "description": "Por que esta mudando. Uma troca sem motivo escrito e a que ninguem consegue explicar tres meses depois."
          }
        },
        "type": "object",
        "required": [
          "plan",
          "reason"
        ],
        "title": "AdminPlanChangeRequest"
      },
      "AdminStatusChangeRequest": {
        "properties": {
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "reason": {
            "type": "string",
            "maxLength": 500,
            "minLength": 5,
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "is_active",
          "reason"
        ],
        "title": "AdminStatusChangeRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}